% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCallsCoMagic.R
\name{getCallsCoMagic}
\alias{getCallsCoMagic}
\title{Get Calls}
\usage{
getCallsCoMagic(token = NULL, user_id = NULL, date_from = NULL,
  date_till = NULL, filter = list(field = "campaign_name", operator =
  "!=", value = NULL), fields = list("id", "visitor_id", "ua_client_id",
  "start_time", "finish_time", "finish_reason", "direction", "is_lost",
  "communication_number", "communication_id", "communication_type",
  "total_duration", "virtual_phone_number", "sale_date", "sale_cost",
  "search_query", "search_engine", "referrer_domain", "referrer",
  "entrance_page", "channel", "tags", "scenario_name", "site_domain_name",
  "campaign_name", "visitor_city", "visitor_region", "visitor_country",
  "visitor_device", "contact_phone_number", "utm_source",     
  "utm_medium", "utm_term", "utm_content", "utm_campaign", "attributes",
  "source"))
}
\arguments{
\item{token}{Your Data API token}

\item{user_id}{Your ID, request it from technical support}

\item{date_from}{date_from, by default adding 00:00:00}

\item{date_till}{date_till, by default adding 00:00:00, so that you should do date+1}

\item{filter}{filters, default value is !=NULL, https://www.comagic.ru/support/api/data-api/#_filters}

\item{fields}{fields, (see \code{\link[fields]{https://www.comagic.ru/support/api/data-api/Reports/}})}
}
\description{
Function to get all calls
}
\examples{
getCallsCoMagic()
}
